fix(CheckboxRadioBase): Resolve 'nested-interactive' axe v4 error
This resolves the following error:

https://dequeuniversity.com/rules/axe/4.4/nested-interactive

This was occurring as we effectively had a checkbox within a checkbox, as we had a visually-hidden `<input>` contained within an element with `role="checkbox"`.

The removes the roles from the parent elements.

A further problem was that the 0px <input> caused a screen reader used with Firefox to behave oddly. This uses a slightly different approach and makes it a child of `StyledCheckmark` with the same dimensions. This required some changes to the CSS.

Finally, a bug where tabbing to a no-container radio caused it to appear checked was also fixed.